# Add a Bitmap to record the tuple id range of all visible tuples