--- a/src/backend/access/appendonly/appendonly_visimap_entry.c
+++ b/src/backend/access/appendonly/appendonly_visimap_entry.c
@@ -450,10 +450,11 @@ AppendOnlyVisimapEntry_GetFirstRowNum(
 bool
 AppendOnlyVisimapEntry_IsVisible(
 								 AppendOnlyVisimapEntry *visiMapEntry,
-								 AOTupleId *tupleId)
+								 AOTupleId *tupleId, Bitmapset **allVisibleSet)
 {
 	int64		rowNum,
-				rowNumOffset;
+				rowNumOffset,
+				rangeNum;
 	bool		visibilityBit;
 
 	Assert(visiMapEntry);
@@ -473,6 +474,8 @@ AppendOnlyVisimapEntry_IsVisible(
 			   "Append-only visi map entry: All entries are visibile: "
 			   "(firstRowNum, rowNum) = (" INT64_FORMAT ", " INT64_FORMAT ")",
 			   visiMapEntry->firstRowNum, rowNum);
+		rangeNum = rowNum / APPENDONLY_VISIMAP_MAX_RANGE;
+		*allVisibleSet = bms_add_member(*allVisibleSet, rangeNum);
 		return true;
 	}
 	Assert(rowNum >= visiMapEntry->firstRowNum);