--- a/src/include/access/appendonly_visimap.h
+++ b/src/include/access/appendonly_visimap.h
@@ -63,6 +63,7 @@ typedef struct AppendOnlyVisimap
 	 * Support operations to search, load, and store visibility map entries.
 	 */
 	AppendOnlyVisimapStore visimapStore;
+	Bitmapset *allvisible_bitmap[AOTupleId_MultiplierSegmentFileNum];
 
 } AppendOnlyVisimap;