--- a/src/include/access/appendonly_visimap_entry.h
+++ b/src/include/access/appendonly_visimap_entry.h
@@ -122,7 +122,7 @@ bool AppendOnlyVisimapEntry_CoversTuple(
 bool AppendOnlyVisimapEntry_IsVisible(
 								 AppendOnlyVisimapEntry *visiMapEntry,
 								 AOTupleId *aoTupleId,
-								 Bitmapset **allVisibleSet);
+								 bool *inAllVisible);
 
 TM_Result AppendOnlyVisimapEntry_HideTuple(
 								 AppendOnlyVisimapEntry *visiMapEntry,