This is intended to add a Bitmap to record the range of AOTupleId per segfile for all visible tuples. The range number is computed as AOTupleId.rowNum / APPENDONLY_VISIMAP_MAX_RANGE indicating the AOTupleId falls into this range, so it could return "visible" directly instead of scanning the visimap table.

Basically, for the sequential indexonlyscan in condition "correlation == 1", the performance improvement is < 10%, for random indexonlyscan in condition "correlation != 1", the performance improvement is > 10%.

Using a simple case to show a slightly enhancement with above optimization:
case1: correlation == 1
```SQL
create table aoco1 (a int, b int, c int) with (appendonly=true, orientation=column, compresstype=zlib, compresslevel=3);
insert into aoco1 select i,i,i from generate_series(1,1000000000)i;
create index on aoco1(a);
vacuum analyze aoco1;
select attname,correlation from pg_stats where tablename='aoco1';
 attname | correlation 
---------+-------------
 a       |           1
 b       |           1
 c       |           1
(3 rows)

gpadmin=#

**Utility Mode on Segment**

[with patch]

gpadmin=# explain analyze select count(*) from aoco1 where a > 991243 and a < 10457620;
                                                                  QUERY PLAN                                                                   
-----------------------------------------------------------------------------------------------------------------------------------------------
 Aggregate  (cost=67611.15..67611.16 rows=1 width=8) (actual time=570.711..570.711 rows=1 loops=1)
   ->  Index Only Scan using aoco1_a_idx on aoco1  (cost=0.44..65255.42 rows=942292 width=0) (actual time=0.131..422.958 rows=3156029 loops=1)
         Index Cond: ((a > 991243) AND (a < 10457620))
         Heap Fetches: 0
 Optimizer: Postgres-based planner
 Planning Time: 2.312 ms
   (slice0)    Executor memory: 1188K bytes (seg0).
 Memory used:  0kB
 Execution Time: 570.744 ms
(9 rows)

Time: 573.390 ms
gpadmin=# 

[without patch]

gpadmin=# explain analyze select count(*) from aoco1 where a > 991243 and a < 10457620;
                                                                  QUERY PLAN                                                                   
-----------------------------------------------------------------------------------------------------------------------------------------------
 Aggregate  (cost=67611.15..67611.16 rows=1 width=8) (actual time=588.552..588.553 rows=1 loops=1)
   ->  Index Only Scan using aoco1_a_idx on aoco1  (cost=0.44..65255.42 rows=942292 width=0) (actual time=0.103..440.983 rows=3156029 loops=1)
         Index Cond: ((a > 991243) AND (a < 10457620))
         Heap Fetches: 0
 Optimizer: Postgres-based planner
 Planning Time: 1.618 ms
   (slice0)    Executor memory: 1188K bytes (seg0).
 Memory used:  0kB
 Execution Time: 588.588 ms
(9 rows)

Time: 590.567 ms
gpadmin=# 
```
case2: correlation != 1
```SQL
create table aoco3 (a int, b int, c int) with (appendonly=true, orientation=column, compresstype=zlib, compresslevel=3);
insert into aoco3 select i,i,i from generate_series(1,1000000)i;
insert into aoco3 select * from aoco3;
insert into aoco3 select * from aoco3;
insert into aoco3 select * from aoco3;
insert into aoco3 select * from aoco3;
insert into aoco3 select * from aoco3;
insert into aoco3 select * from aoco3;
insert into aoco3 select * from aoco3;

select count(*) from aoco3;
create index on aoco3(a);
vacuum analyze aoco3;
select attname,correlation from pg_stats where tablename='aoco3';
 attname | correlation 
---------+-------------
 a       | 0.005129206
 b       | 0.005129206
 c       | 0.005129206
(3 rows)

gpadmin=# set enable_bitmapscan to off;
SET
Time: 0.180 ms

[with patch]

gpadmin=# explain analyze select count(*) from aoco3 where a > 991243 and a < 10457620;
                                                                  QUERY PLAN                                                                  
----------------------------------------------------------------------------------------------------------------------------------------------
 Aggregate  (cost=48130.88..48130.89 rows=1 width=8) (actual time=259.272..259.272 rows=1 loops=1)
   ->  Index Only Scan using aoco3_a_idx on aoco3  (cost=0.43..47799.53 rows=132539 width=0) (actual time=0.141..240.280 rows=367488 loops=1)
         Index Cond: ((a > 991243) AND (a < 10457620))
         Heap Fetches: 0
 Optimizer: Postgres-based planner
 Planning Time: 2.345 ms
   (slice0)    Executor memory: 1188K bytes (seg0).
 Memory used:  0kB
 Execution Time: 259.304 ms
(9 rows)

Time: 261.985 ms
gpadmin=#

[without patch]

gpadmin=# explain analyze select count(*) from aoco3 where a > 991243 and a < 10457620;
                                                                  QUERY PLAN                                                                  
----------------------------------------------------------------------------------------------------------------------------------------------
 Aggregate  (cost=48130.88..48130.89 rows=1 width=8) (actual time=503.359..503.360 rows=1 loops=1)
   ->  Index Only Scan using aoco3_a_idx on aoco3  (cost=0.43..47799.53 rows=132539 width=0) (actual time=0.105..483.132 rows=367488 loops=1)
         Index Cond: ((a > 991243) AND (a < 10457620))
         Heap Fetches: 0
 Optimizer: Postgres-based planner
 Planning Time: 1.127 ms
   (slice0)    Executor memory: 1188K bytes (seg0).
 Memory used:  0kB
 Execution Time: 503.392 ms
(9 rows)

Time: 504.847 ms
gpadmin=#
```
case3: memory usage comparison
```SQL
create table aocopart (i int, j int) using ao_column partition by range(j) (start(1) end(1000001) every(1000));
insert into aocopart select 2, generate_series(1,1000000);
select count(*) from aocopart;
create index on aocopart(j);
vacuum analyze aocopart;
select attname,correlation from pg_stats where tablename='aocopart';
gpadmin=# explain analyze select count(*) from aocopart where j > 991243 and j < 10457620;
                                                                 QUERY PLAN                                                                 
--------------------------------------------------------------------------------------------------------------------------------------------
 Finalize Aggregate  (cost=75.95..75.96 rows=1 width=8) (actual time=125.808..125.810 rows=1 loops=1)
   ->  Gather Motion 3:1  (slice1; segments: 3)  (cost=75.89..75.94 rows=3 width=8) (actual time=61.787..125.788 rows=3 loops=1)
         ->  Partial Aggregate  (cost=75.89..75.90 rows=1 width=8) (actual time=97.537..97.541 rows=1 loops=1)
               ->  Append  (cost=0.00..68.59 rows=2918 width=0) (actual time=11.039..97.086 rows=8757 loops=1)
                     ->  Seq Scan on aocopart_1_prt_992  (cost=0.00..6.00 rows=252 width=0) (actual time=11.038..11.170 rows=757 loops=1)
                           Filter: ((j > 991243) AND (j < 10457620))
                           Rows Removed by Filter: 243
                     ->  Seq Scan on aocopart_1_prt_993  (cost=0.00..6.00 rows=333 width=0) (actual time=10.778..10.935 rows=1000 loops=1)
                           Filter: ((j > 991243) AND (j < 10457620))
                     ->  Seq Scan on aocopart_1_prt_994  (cost=0.00..6.00 rows=333 width=0) (actual time=10.816..10.972 rows=1000 loops=1)
                           Filter: ((j > 991243) AND (j < 10457620))
                     ->  Seq Scan on aocopart_1_prt_995  (cost=0.00..6.00 rows=333 width=0) (actual time=10.555..10.713 rows=1000 loops=1)
                           Filter: ((j > 991243) AND (j < 10457620))
                     ->  Seq Scan on aocopart_1_prt_996  (cost=0.00..6.00 rows=333 width=0) (actual time=10.435..10.595 rows=1000 loops=1)
                           Filter: ((j > 991243) AND (j < 10457620))
                     ->  Seq Scan on aocopart_1_prt_997  (cost=0.00..6.00 rows=333 width=0) (actual time=10.388..10.546 rows=1000 loops=1)
                           Filter: ((j > 991243) AND (j < 10457620))
                     ->  Seq Scan on aocopart_1_prt_998  (cost=0.00..6.00 rows=333 width=0) (actual time=10.379..10.534 rows=1000 loops=1)
                           Filter: ((j > 991243) AND (j < 10457620))
                     ->  Seq Scan on aocopart_1_prt_999  (cost=0.00..6.00 rows=333 width=0) (actual time=10.406..10.563 rows=1000 loops=1)
                           Filter: ((j > 991243) AND (j < 10457620))
                     ->  Seq Scan on aocopart_1_prt_1000  (cost=0.00..6.00 rows=333 width=0) (actual time=10.343..10.500 rows=1000 loops=1)
                           Filter: ((j > 991243) AND (j < 10457620))
 Optimizer: Postgres-based planner
 Planning Time: 179.583 ms
   (slice0)    Executor memory: 79K bytes.
   (slice1)    Executor memory: 4047K bytes avg x 3 workers, 10260K bytes max (seg0).
 Memory used:  128000kB
 Execution Time: 147.270 ms
(29 rows)

Time: 332.456 ms
gpadmin=#
```
![image](https://github.com/greenplum-db/gpdb/assets/90174703/3064b665-31e5-456e-af68-c9fe7c7622e3)

The red color indicates the fixed size of difference introduced by the patch (because of the memory allocation of 2D array), which is +1MB per physical table scan (under scan descriptor's lifecycle). The new extra 1MB memory won't be cumulated as long as the table scan descriptor is destroyed.  For sure there is also variable size of increasing of memory usage which is depending on the number of tuples to be scanned.

## Here are some reminders before you submit the pull request
- [ ] Add tests for the change
- [ ] Document changes
- [ ] Communicate in the mailing list if needed
- [ ] Pass `make installcheck`
- [ ] Review a PR in return to support the community