@haolinw 2023-11-07
> Some thoughts:
> 
> 1. About the cost. As discussed, the max memory requirement is 512MB per table, but it's just the worst case. It's not easy to calculate the cost accurately because it heavily depends on the table size, the data layout and the query itself. For a complex query involving a couple of large tables, I suppose it's possible to require hundreds of MBs. In addition, **the memory requirement is not accounted to resource management (RQ or RG)**. It means we might got incorrect memory calculation or even OOM.
> 2. About the benefit. Again, it's not easy to calculate the benefit accurately. As my understanding, the benefit is max when there is almost no IUD on the table (i.e. all tuples are visible), and we access the same tuples for many times (so the statistic info works). Maybe we can do more test to hit the conditions and get more performance improvement.
> 3. `visiMap->allvisibleset` will be freed when the query closes, which means the statistic info must be rebuilt for next query. It's the most significant disadvantage. Is it possible to reuse the info between queries? I know it's difficult, but it might be feasible via kind of buffer mechanism (beyond this PR).
> 4. As discussed, did you ever test the code on 32bit env to check 32bit bms behavior? I suppose it would work, but it's safer to run a pipeline.
> 5. Since both cost/benefit are not so predictable, I'd suggest to keep `gp_aovisimap_opt` permanent so we will have a knob to switch the behavior at least.
> 
> Let's talk offline.

Sure. Also include @ashwinstar and @soumyadeep2007 for above discussion in case of any different thought for 2D array bitmapset approach to address aovisimap caching.