@haolinw 2023-11-07
> 1. In addition, **the memory requirement is not accounted to resource management (RQ or RG)**. It means we might got incorrect memory calculation or even OOM.

Do you have any real case to explain this ? Because all the optimization works under `VisiMapContext` so don't see why it can't be controlled well by RQ/RG.