@yaowangm 2023-11-07
I had a discussion with @haolinw. Regardless of all other details, the major concern is about memory cost.

When I mentioned "the memory requirement is not accounted to resource management (RQ or RG)", I meant the code doesn't consider the memory quota (upper limit) of an executor. For example, a hash join executor will create hash table according to the the memory quota, and when the memory usage is near to the quota, it will start to spill the data to disk to ensure the quota is obeyed. But for this code, the memory requirement will keep increasing to over the quota (up to 512MB per table) and hit OOM finally. As our analysis, it's very possible for complex query/large tables and we will break many queries which work well previously.

It seems spilling is not an option for the case, and I agree that we don't need to impose too much complexity here. What I can imagine now is:

1. In `AppendOnlyVisimapAllVisibleSet_AddTuple()`, check if we are near to the memory quota. If yes, cancel adding more bms.
2. Keep `gp_aovisimap_opt` permanent so we will have a workaround to disable the mechanism at least when customer hit OOM.

@ashwinstar @soumyadeep2007 Please give your idea if necessary, thanks!