@haolinw 2023-11-08
Verified both 64bit and 32bit bitmapword passed all pipeline tests.