@haolinw 2023-11-09
> 2. Keep gp_aovisimap_opt permanent so we will have a workaround to disable the mechanism at least when customer hit OOM.

Agree on this, renamed `gp_aovisimap_opt` to `gp_enable_aovisimap_allvisibleset` to make it permanently.  Also think keeping this GUC may help to better identify the effectivity of this optimization and the pattern of aovisimap scan in various real cases, test in dev is too limit to form a solid model.