@haolinw 2023-11-15
> 1. In AppendOnlyVisimapAllVisibleSet_AddTuple(), check if we are near to the memory quota. If yes, cancel adding more bms.

An ideal approach is having `palloc` and `repalloc` return NULL quietly just like `palloc_extended(size, MCXT_ALLOC_NO_OOM | MCXT_ALLOC_ZERO)`, and being called by bms_add_member and checked the result from upper caller (`AppendOnlyVisimapAllVisibleSet_AddTuple`), if NULL pointer is returned, stop growing (or allocating) the bitmapset array hence keep allvisibleset serving those rangenums which are already contained.

But this approach needs deep upstream code injection (or duplication), and don't think other approach like predicating a threshold to control the behavior is good enough to solve this problem, so won't do that unless reaching consensus or having other better alternative.