@yaowangm 2023-11-15
> > 1. In AppendOnlyVisimapAllVisibleSet_AddTuple(), check if we are near to the memory quota. If yes, cancel adding more bms.
> 
> An ideal approach is having `palloc` and `repalloc` return NULL quietly just like `palloc_extended(size, MCXT_ALLOC_NO_OOM | MCXT_ALLOC_ZERO)`, and being called by bms_add_member and checked the result from upper caller (`AppendOnlyVisimapAllVisibleSet_AddTuple`), if NULL pointer is returned, stop growing (or allocating) the bitmapset array hence keep allvisibleset serving those rangenums which are already contained.
> 
> But this approach needs deep upstream code injection (or duplication), and don't think other approach like predicating a threshold to control the behavior is good enough to solve this problem, so won't take action unless reaching consensus or having other better alternative.

Maybe we don't need to make that trouble. A rough idea is, in `AppendOnlyVisimapAllVisibleSet_AddTuple()`, simply check the memory required and stop growing bitmapset if the limit is hit.

Some pseudo code:
```
// add a members isMemFull to AppendOnlyVisimapAllVisibleSet
// We can find mem_used/mem_limit for an executor from context

// In AppendOnlyVisimapAllVisibleSet_AddTuple()
// calculate the delta memory
long deltaMem = (WORDNUM(x) - *bmspp->nwords) * SIZEOF_VOID_P;
if(mem_limit + deltaMem > mem_limit)
{
  isMemFull = true;
}else{
  *bmspp = bms_add_member(*bmspp, rangenum); 
}
```