@haolinw 2023-11-16
> mem_limit

How can we get `mem_limit` ? This is what I don't like introducing as you have to pass such variable all the way to storage objects or interfaces which should be not responsible for the requirement.